# plugins/ui-rest/openapi.yaml
openapi: 3.1.0
info:
  title: HiveGuard ui.rest API
  version: "1.0.0"
  description: |
    HTTP + WebSocket management and telemetry API exposed by the `ui.rest`
    plugin of the HiveGuard daemon. This is the consolidated REST surface
    (legacy `api.*` + `ui.rest` merged — see API_REFACTOR.md).

    ## Authentication
    All endpoints under `/api/*` require a bearer token **except** the public
    endpoints `GET /api/health`, `GET /api/stream` (WS) and `GET /metrics`.
    Present it as `Authorization: Bearer <auth_token>` where `auth_token` is
    the value configured for the plugin.

    ## CORS
    The server emits CORS headers only for the origins listed in the plugin's
    `cors_origins` config. A standalone frontend (see `../hiveguard-web/`) must
    have its origin whitelisted there, or be served same-origin via `static_dir`.

    ## Schema source of truth
    Response/request shapes mirror the Rust wire types in
    `hiveguard-plugin-api/src/traits/ui_server.rs` 1:1.
servers:
  - url: http://127.0.0.1:8443
    description: Default local bind (see plugin `bind_addr`).
security:
  - bearerAuth: []
...
paths:
...
  /api/ingest/logs:
    post:
      tags: [telemetry]
      summary: HTTP-push log ingest (mounted only when ingest is enabled)
      description: |
        Carries its own bearer gate (`ingest.token`, falling back to
        `auth_token`), a per-second rate limit, and a body-size cap.
      operationId: ingestLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Log line(s) in the daemon's accepted ingest format.
              type: object
      responses:
        "202": { description: Accepted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { description: Payload too large }
        "429": { description: Rate limited }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Value of the plugin's `auth_token`.
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...